requirements.txt
absl-py
pytest
sacrebleu==1.4.10
six
wheel
progressbar
sklearn
torch==1.6.0 
torchvision==0.7.0
transformers==3.1.0
pyyaml
nltk
jupyter
notebook
